+++ Entity/Enemy/Enemy.h
@@ -16,15 +16,15 @@ public:
     void setHealth( int health );
     int getHealth() const;
 
-    void setSpeed( int speed );
-    int getSpeed() const;
+    void setSpeed( float speed );
+    float getSpeed() const;
 
     void update( sf::Time& deltaTime );
     void render( sf::RenderWindow& window );
 
 private:
     int _health;
-    int _speed;
+    float _speed;
 };
 
 END_ENTITY_NAMESPACE
